fix: Ensure verifyKvStoreConnection takes into account rootPrefix
Motivation

Instances of KVUtilsFactory are initialized for a particular connection configuration, which for etcd includes a "chroot" rootPrefix. All the utility methods honor this prefix, except the verifyKvConnection method.

This can cause problems when role-based access controls are configured in etcd and the client-configured userid doesn't have access outside of the rootPrefix key range. In this case the verifyKvConnection call will fail unexpectedly with a permission denied error.

Modifications

- Adjust the verifyKvConnection to perform a get on a dummy (likely nonexistent) key within the rootPrefix range
- Add auth-based unit test

Result

KVUtilsFactory#verifyKvConnection will work as expected in key-specific auth cases.

Signed-off-by: Nick Hill <nickhill@us.ibm.com>